Rename main component
